Complete #30 New tests for required option
